+++ logger_test.go
@@ -114,26 +114,6 @@ func TestLoggerFilterLog(t *testing.T) {
 	})
 }
 
-func TestLoggerLogInvalidJSON(t *testing.T) {
-	test.WithLogger(t, func(logger *xylog.Logger, w *test.MockWriter) {
-		logger.Handlers()[0].SetFormatter(xylog.NewJSONFormatter())
-		logger.Event("test").Field("func", func() {}).Error()
-		xycond.ExpectIn("An error occurred while formatting the message",
-			w.Captured).Test(t)
-	})
-}
-
-func TestLoggerAddExtraMacro(t *testing.T) {
-	test.WithLogger(t, func(logger *xylog.Logger, w *test.MockWriter) {
-		var formatter = xylog.NewTextFormatter().AddMacro("foo", "custom")
-		logger.AddExtraMacro("custom", "this is a custom field")
-		logger.Handlers()[0].SetFormatter(formatter)
-		logger.Event("test").Error()
-		xycond.ExpectIn(`foo="this is a custom field" event=test`,
-			w.Captured).Test(t)
-	})
-}
-
 func TestLoggerAddField(t *testing.T) {
 	test.WithLogger(t, func(logger *xylog.Logger, w *test.MockWriter) {
 		logger.AddField("custom", "this is a custom field")
@@ -165,4 +145,21 @@ func TestLoggerFilters(t *testing.T) {
 
 	logger.RemoveFilter(filter)
 	xycond.ExpectEqual(len(logger.Filters()), 0).Test(t)
+}
+
+func TestLoggerFindCaller(t *testing.T) {
+	xylog.SetFindCaller(true)
+	defer xylog.SetFindCaller(false)
+	test.WithLogger(t, func(logger *xylog.Logger, w *test.MockWriter) {
+		var handler = logger.Handlers()[0]
+		handler.AddMacro("module", "module")
+		handler.AddMacro("funcname", "funcname")
+
+		logger.Error("foo")
+
+		xycond.ExpectIn(
+			"module=github.com/xybor-x/xylog_test.TestLoggerFindCaller",
+			w.Captured).Test(t)
+		xycond.ExpectIn("funcname=func1", w.Captured).Test(t)
+	})
 }